Add Singleton and Config Manager classes
SingletonCxx - A superclass that should be inherited from when creating
singletons that will only be used within C++

SingletonCxxQml - A superclass that should be inherited from when creating
singletons that will appear in both C++ and QML code.

ConfigManager - A SingletonCxxQml that holds a map of configurations that are
read in via a specified config file.

Type: New Feature